Upload of Beckhoff Automation and TwinCAT3 logo
We give the approval to use the uploaded logos of Beckhoff and TwinCAT3 on onnx.ai website as well as in presentations showing ONNX Member Companies.
The rights to use the Beckhoff and the TwinCAT 3 logo are restricted to the usage approved by us.

The Beckhoff logo could be placed on onnx.ai page, while TwinCAT3 logo could be placed on supported Tools (Frameworks and converters). Please link to https://www.beckhoff.com/machine-learning/

Thanks!