The attached Abbyy logo can be used on the onnx.ai website as well as in presentations showing ONNX Member Companies. It can be linked to the following url: https://www.abbyy.com/neoml/
The attached NeoML logo can be used on the onnx.ai website as well as in presentations showing ONNX Member Companies.
It can be linked to the following url: https://github.com/neoml-lib/neoml

ABBYY_logo.png is a company logo for main page.
NeoML_logo.png is a tool logo for supported_tools/inference page.

Signed-off-by: Stanislav Angeliuk <stanislav.angelyuk@abbyy.com>